fix(sdk): inflight clients are not cached (#5478)
Reuse the handler's inflight client across `cloud.Function` invocations, as long as they are running in the same sandbox.

This enables common FaaS optimizations such as caching database connections, etc.

To test this and improve the fidelity of our simulator, I've changed the behavior of `cloud.Function` in the simulator to reuse the function sandbox across invocations and just call `handle()` every time.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [x] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.