feat(sdk): bundle functions eagerly in simulator (#5499)
Fixes #5031

Sometimes [laziness is a virtue](https://github.com/winglang/wing/pull/5453). Other times (like now), eagerness is a virtue. It just depends on the situation. 🤓 

To improve interactivity in the Wing Console and Wing Playground, this PR updates the local simulator so that it eagerly bundles all functions while starting a simulation, instead of the previous behavior of waiting until the first time a function is invoked. This means that in most cases, by the time you try interacting with any `cloud.Function`, the code will probably have finished bundling in the background, and so clicking "Invoke" will only take as much time as it takes to actually execute the inflight code.

Since bundled code can have side effects at the top-level (for example, any code that runs inside `inflight new() {}` blocks), we still only load the bundled code into the VM upon the first time a function invocation is made, to guarantee that you get FaaS cold-start-like behavior. I've updated one of our tests (inflight_handler_singleton.test.w) to demonstrate this.

If you happen to click on the "invoke" button on a function before it has finished bundling, it's possible you might still have to wait a little bit extra - but even then it should be a speedup. 🏃

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.