feat(console): allow dragging the map with the mouse
Devices with touchpads can move the map around with the usual scrolling gestures (typically, using two fingers), but devices with keyboard and mouse need an alternative way to easily move the map around. This feature allows clicking on the map with a mouse and moving it around, in addition to the recent two-finger gesture.

Solves #5446.