chore: turbo diff check failing for transitive dependencies (#5486)
This script is responsible for checking what projects/tasks changed in CI. There are cases where certain transitive dependencies don't bubble up all the way. Importantly, this caused the e2e to be ignored in a PR like this https://github.com/winglang/wing/pull/5484

This change now very explicitly lists out and checks transitive deps. Honestly it's pretty naive but at the scale we're dealing with it's not a big deal.

The break may be a result of updating turbo. The list returned from the turbo cli previously always seemed to be in topological order, and it wasn't for the above PR.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.