chore: compatibility-spy build fails on PRs from forks (#5528)
The recent addition of the compatibility-spy lib broke fork workflows. This is because it was attempting to compile during the e2e matrix, but it fails because the e2e matrix is not actually meant to rebuild things. This accidentally works on PRs from winglang/wing because we use turbo remote caching, so the `turbo compile` command pulled from the remote cache.

This PR instead utilizes @winglang/compatibility-spy as a package in hangar, avoiding any re-compilation. While there I also cleaned up the tsconfig and changed it to use tsup (since it was installed anyways).

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.