fix(compiler)!: interface methods can be declared with no return type (#5522)
Fixes #5454 by raising an error when interface methods are declared without return types.

In addition to requiring return types on interfaces, I also made them required on extern function statements.

The language reference is a ambiguous and doesn't explicitly say return types are optional on class methods. It's probably a good idea to require return types there as well, but I imagine this is more controversial since some might argue it's okay to omit them and let the compiler infer them from function bodies. To avoid controversy I've left the existing behavior as is.

BREAKING CHANGE: Interface methods now require return types. This was required in earlier versions of Wing (v0.18.x) but due to regressions it was not enforced, so out of caution this change has been labeled as potentially breaking.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.