chore: switch to Node LTS (v20) for internal dev and recommend it for users (#5580)
This switches the default for dev to be Node20.

Wing itself should still work with 18 and I guess we can continue doing e2e tests for it for now, but the docs for users has also changed to recommend v20.

Updated pnpm and TSX because I got a couple new warnings, updating fixed it.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.