fix: ts types for optional requires explicit `undefined` (#5604)
wingc currently emits an optional type in functions/fields like this:

```ts
interface StructyGuy {
  a: (str) | undefined
}
```

While this is technically correct for the purpose of the runtime, it doesn't allow you to completely omit `a`.
This PR generates it like this:

```ts
interface StructyGuy {
  a?: (str) | undefined
}
```

This may seem verbose, but in wing the `?` on a type technically does mean both of these things. It's possible that in certain contexts we could simplify it to improve how it looks to users, but this behavior is still correct so I think it's fine to continue on this path.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.