fix(compiler): onLift not called, or called too many times (#5593)
This PR refactors code defining the interface between the compiler and core TypeScript library, the pieces of which are responsible for "lifting" objects and granting capabilities (like environment variables and IAM permissions) through the `onLift` and `onLiftType` methods. Through this refactoring we're able to fix a couple regressions that came out of #5435.

Today, the contract we’re trying to provide with the public `onLift()` method which can be added to any preflight class, is as follows. `onLift()` will be called on an object at most per inflight `host` argument.

As a concrete example, suppose you write some code like this in your class:

```js
// ... 
class A {
  onLift(host: std.IInflightHost, ops: Array<str>) {
    log(“I was lifted by “ + host.node.path);
  }
}
// …
```

Then you should never see the same line printed two or more times:
```
$ wing compile main.w
I was lifted by root/cloud.Function
I was lifted by root/cloud.Function
```

The motivation is this that it’s a better DX for library authors. The main utility of `onLift` is to add mutations or additions to the `host`. The idempotency requirement ensures you don’t have to defensively code as if the method will be called multiple times with the same host.

The implementation of this has been tricky though for various reasons (like lazy execution, and the fact that we have to deal with class inheritance). In part, #5435 introduce a regression where this contract was violated.

At a high level, we've fixed this by replacing an internal utility method, `onLiftMatrix()`, (which partially de-duplicated `onLift` calls) with a separate algorithm that traverses an object dependency graph based the information we're generating with the compiler. The core logic is implemented in lifting.ts (https://github.com/winglang/wing/blob/43d7864d3509dd08085da2835dfacf36783318d2/libs/wingsdk/src/core/lifting.ts).
A collection of objects and each of the operations requested on them is generated. Once we have this collection, only then do we call `onLift` and `onLiftType` methods.

Several tests have been added to validate edge cases where de-duplication issues arise.

Fixes #5561
Fixes #5573

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.