feat(sdk)!: improve `api` and `function` handlers
This change introduces a few ergonomic improvements to the inflight handlers accepted by `cloud.Api` and `cloud.Function`.

The signature of the inflight closure accepted by`cloud.Function` is now:

    (event: str?): str?

Both the input and the output as now optional. The `invoke()` and `invokeAsync()` operations were updated accordingly.

The handlers accepted by `cloud.Api` now return an optional `ApiResponse?` and the `status` field in `ApiResponse` is now optional as well and defaults to 200.

This allows the handler either to return just a `body` or not return anything. In body cases the result is 200 OK.

    api.get("/", inflight () => {
      return { body: "hello" };
    });

    api.post("/cows", inflight () => {
      log("cow added");
    });

BREAKING CHANGE: The `function.invoke()` and `function.invokeAsync()` methods now return `str?` instead of `str`.