feat(compiler): bang operator (#5556)
I added the bang operator as described in #4923.

`throw` is a statement, so `x!` cannot be compiled to: `x ?? throw 'nil'`, It is also not enough to use block: `x ?? { throw 'nil' }` (syntax error). The only way is to use a function: `x ?? (() => throw 'nil')()`, and that's how I implemented it.

More precisely, `x!` becoming:

```
(x??(()=>{throw new Error("Unexpected nil");})())
```