fix(compiler)!: default object ids are volatile because of namespace
Use only the class name instead of the fully qualified name as the default preflight object ID because the namespace (e.g. `foo` in `foo.MyClass`)
is a local name determined by the `bring` statement.

Fixes #5564

BREAKING CHANGE: caution (and apologies)! This change is going to cause a replacement of any resource that didn't have an explicit identity.
We are working to make it [safer and easier](https://github.com/winglang/wing/issues/901) to control IDs of resources (especially ones with state).