feat(compiler): bypass throw at parent level and disregard return type (#5555)
Based on #5414 

## Description

Added a bypass check at the return type level when a throw is detected in the init statement. In addition, I implemented a more generic visitor iterator that allows us to visit and collect not only the return type of the throw type and potentially more statements. 

Before:
```
error: A function whose return type is "num" must return a value.
  --> wing/main.w:1:24
  |  
1 |   let x: (): num = () => {
  | /------------------------^
2 | |   throw "not implemented";
3 | | };
  | \-^ A function whose return type is "num" must return a value
```

Now this code executed well and it has also tests to support this claim

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.