back to lifting all preflight expressions referencing inflight closures
Because the type system can't distinguish between a closure class (which needs to be lifted) and a regular function type...
Also lift on reference and not all call, by assuming a ref to a closure will actually call it later on, we can correctly lift (and setup permissions) for closures stored in collections (for example). This is a special exception to our qualification rules, since closures only have a single method (`handle`) so we can implicitly qualify it.