feat(sdk)!: improve api and function handlers (#5644)
This change introduces a few ergonomic improvements to the inflight handlers accepted by `cloud.Api` and `cloud.Function`.

### `cloud.Function`

The signature of the inflight closure accepted by`cloud.Function` is now:

```js
(event: str?): str?
```

Both the input and the output as now optional. The `invoke()` and `invokeAsync()` operations were updated accordingly.

### `cloud.Api`

The handlers accepted by `cloud.Api` now return an optional `ApiResponse?` and the `status` field in `ApiResponse` is now optional as well and defaults to 200.

This allows the handler either to return just a `body` or not return anything. In body cases the result is 200 OK.

```js
api.get("/", inflight () => {
  return { body: "hello" };
});

api.post("/cows", inflight () => {
  log("cow added");
});
```

BREAKING CHANGE: The `function.invoke()` and `function.invokeAsync()` methods now return `str?` instead of `str`.
* **sdk:** The `event` argument accepted by `cloud.Function` handlers is now `str?` instead of `str`.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [x] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.