feat(sdk): implement `entries()` method inside Map (#5475)
Related to #5005

## Description

Fixes #5005, we wanted to add the `entries()` method to `Map` instances (*and `MutMap`). This essentially returns an iterator and from that, via a macro, an array of the form [{ key, value }] for each element within the `Map`.

```ts
let map = { "foo" => "hello" };

for x in map.entries() {
  log(x.key);
  log(x.value);
}
```

Now returns 

> foo
> hello

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.