fix(compiler): closures in collections not lifted or not qualified correctly (#5757)
See #5527
Collections containing inflight closures now implicitly lift and qualify all the closures with them with their generated `handle` method.
This PR extends the existing assumption done by the compiler where any reference to an inflight closure means it'll be called so we should qualify it with `handle`, this assumption is now extended to any inflight closure in a lifted collection.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.