feat(compiler)!: create sets with `[]` instead of `{}` (#5729)
Based on #4843 

## Description

Implemented new syntaxis for the `Set` and `MutSet` keywords. From now on, sets will be defined using the identical syntaxes used for Arrays with the difference that there won't be any implicit definition of a set (before it was plain `{}`)

Example

```rust
let set : Set<num> = Set<num> [1,2,3];
```

I limited myself to adding more tests since this was more a functionality related to all the tests we already had in place.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.

BREAKING CHANGE: `Set` literal syntax (`{1, 2, 3}`) is removed. Use `[]` with an explicit `Set<T>` instead.