feat: add support for displaying different icons based on browser support in endpoint list
The changes in this commit add support for displaying different icons in the endpoint list based on the browser support of each endpoint. If an endpoint has browser support, a globe icon is displayed, and if it doesn't have browser support, a link icon is displayed. This provides visual cues to users about the browser compatibility of each endpoint.