fix: unable to resolve tokens inflight from awscdk platform (#5786)
Fixes #5510

The awscdk platform was importing `registerTokenResolver()` directly from `@winglang/sdk/lib/core/tokens`. This was bypassing preflight logic to always force imports from the SDK to resolve from the same place because it was only checking for imports specifically from `@winglang/sdk` only. This caused it to load it as a different module which had a global that was unused by the rest of the SDK logic.

I was able to replicate the original issue and this fixed it in that same scenario.

Automated testing this is very difficult/slow because it requires a certain scenario that is very realistic: A global install of the SDK and a "local" install of a platform like awscdk. When you install `@winglang/platform-awscdk` it comes with its own copy of `@winglang/sdk`. In our hangar tests this gets deduped because they are installed alongside each other. Adding a way to test this is a pretty big tasks and should at least involve https://github.com/winglang/wing/issues/3390 as well

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.