fix: compilerOptions for typescript project not respected (#5789)
Fixes #5788

There are a few options that are still not supported because it's either necessary (like `outDir`) or difficult (like `paths`). The user will be warned if they provide these options.

Also removed usage of `baseUrl` here because it's not needed and was causing a confusing error in some cases (if you have a file that has the same name as a node_module).

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.