feat(console): restrict dragging the map when using the space key (#5808)
The changes in this commit restrict dragging the map using the space key. When the space key is pressed, the map becomes draggable and the user can move it around. The cursor also changes to a grab cursor when the space key is pressed. This provides a more intuitive way for users to interact with the map.