feat(sdk): in-place app updates in simulator [WIP]
In order to reduce iteration time when updating an application, introduce a new simulator API to support in-place updates of an app in the simulator.

It is now possible to use `simulator.update(simDir)` without needing to stop the simulator and start a new instance. This method will analyze the diff between the currently loaded app and the new app defined in `simDir` and will create/update/delete resources accordingly.

This implies that if a resource was not updated, it won't be restarted across app updates.

As a future improvement, we can add support for watching the sim directory and automatically updating the app inside the simulator. Wow!