chore: mutation workflow and owner comment workflow broken on forks (#5820)
Fixes #5115

The mutation workflow is using the wrong `head` for forks when the repo name is `wing` (which it usually is). This worked before, I suspect the API changed, but in any case this fixes things.

Also updated the initial PR comment workflow to work on forks.

Tested in https://github.com/MarkMcCulloh/wing-distributed-workflow

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.