feat: rust-based CLI (#5829)
Proof of concept for a Rust-based CLI implementation. This PR implements the basics of compiling Wing programs to see how viable this would be - it's missing a lot of other pieces (custom platforms, tests, error handling etc.)

The main thing that's really nice about this is that the contributing experience for the compiler feels a lot better ~You no longer have to run `pnpm turbo compile --filter winglang` to build the CLI (which tends to be slow as it has to build a lot of other dependencies), instead~ since you can run `cargo run` from the CLI directory. You can also run `cargo install --path . --force --locked` to install the binary to your path.

When the CLI is installing the Wing SDK, it gets installed to `~/.wing/cache`, and accessed via NODE_PATH.

https://github.com/winglang/wing/assets/5008987/339ebce1-e38c-4a91-96a4-a85fef66cd90

Help screen:
<img width="589" alt="Screenshot 2024-03-01 at 7 45 50 PM" src="https://github.com/winglang/wing/assets/5008987/1ef90654-35a4-484f-8b48-2a69d2cc97f3">

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.