feat(sdk): cloud.Function concurrency for `sim` target (#5867)
This PR implements a more precise system in the local simulator for simulating inflight function concurrency in cloud.Function, with the goal of unblocking support for function initialization caching (for example, to cache database connections between multiple requests).

Now, the simulation manages a pool of workers, each of which executes inflight code in an isolated Node.js process. Function initialization is performed once on a worker, and can then handle multiple sequential requests thereafter. If a worker is busy handling an existing request, then a new worker is spun up as long as it doesn't exceed the cloud.Function's configured `concurrency`. If a function times out, the worker process has to be stopped, so any initialization state is lost.

This diagram illustrates what could happen when your function receives 10 requests:

![](https://docs.aws.amazon.com/images/lambda/latest/dg/images/concurrency-3-ten-requests.png)

As your function receives more concurrent requests, the simulator scales up the number of execution environment instances in response.

Closes #168
Closes #5864
Closes #5549

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.