feat(sdk)!: change DynamoDB from single table to global table for awscdk target (#5833)
DynamoDB `TableV2` global table approach is now the recommended construct for all new `awscdk` projects, signifying a shift in best practices.

As we haven't configured replicas in the current implementation of `ex.DynamoDBTable`, it doesn't seem necessary to change anything in the `tf-aws` platform. However, in `awscdk`, it's necessary to change the resource to `TableV2`

Closes #5612 

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.