chore: fix nondeterministic bucket sim test (#5896)
Fixes #5895

The goal of the test is to make sure we don't emit events for object creation twice. We can just avoid the problem of stopping the sim by doing it after getting the snapshot fails.

No matter what, the sim will be stopped due to the way the test SimApp works.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.