feat(sdk): in-place app updates in simulator (#5821)
In order to improve iteration speed when updating an application, introduce a new simulator API to support in-place updates of an app in the simulator.

It is now possible to use `simulator.update(simDir)` without needing to stop the simulator and start a new instance. This method will analyze the diff between the currently loaded app and the new app defined in `simDir` and will create/update/delete resources accordingly.

This implies that if a resource was not updated, it won't be restarted across app updates.

The way this works is that when `update(simDir)` is called, we read the new `simulator.json` file and analyze the diff between the current resource configuration and the new one to determine which resources were *added* (exists only in the new config), *removed* (doesn't exist in the new config), *updated* (an attribute was changed) or *retained* (no change). Then, we:

1. Stop all *deleted* and *updated* resources.
2. Copy the state from all the *retained* resources (attributes and resolved properties).
3. Start all *new* and *updated* resources.

Modeled dependencies via an execution graph. This allows us to automatically replace any dependent resources and start in topological order.

Until we can deprecate `sim.State` (waiting for #5882), we sill have an initialization loop which allows late resolution of tokens.

Next step is to update the Wing Console to use this new API when a new version of the app is compiled.

### Future improvements

- Currently all updates are always a replacement (stop and start), but technically we can
  give resources an opportunity to do an in-place update (basically call `resource.update()`).
- We can add support for watching the sim directory and automatically 
  updating the app inside the simulator. Wow!

### Misc

The `api.cleanup()` method did not wait for the server to close.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [x] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.