feat(compiler)!: stringify enums to their own names (#5889)
It's common to want to be able to stringify enums in some way. While the exact mechanism hasn't been defined in the [Wing language spec](https://www.winglang.io/docs/language-reference#38-enumeration) yet, it's technically been possible to stringify enums through a bug in the compiler that allowed any type to be stringified:

```js
bring cloud;

enum Result {
  SUCCESS,
  FAILURE
}

log("{Result.SUCCESS}");
```

Previously this printed out "0". To align with the default behavior of many other languages, this PR updates the behavior so enums stringify by default to their own names (in this case, "SUCCESS").

With that, the PR fixes the gap that allowed all types to be stringified, so for example the following is now an error:

```js
class A {
  privateField: str;
  new() {
    this.privateField = "internal";
  }
}

let a = new A();
log("{a}");
```

This produces the error:

```
error: Expected type to be stringable, but got "A" instead
  --> main.w:9:7
  |
9 | log("{a}");
  |       ^ Expected type to be stringable, but got "A" instead
  |
  = hint: str, num, bool, json, and enums are stringable
```

If you needed this behavior, you can work around it by defining an [extern JavaScript function](https://www.winglang.io/docs/language-reference#52-javascript) that prints the value or calls "toString()", but please note the stringification of these types is subject to change.

This issue tracks the default stringification for more types: https://github.com/winglang/wing/issues/5890

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.