fix: unable to debug wing inflights (#5952)
This specifically fixes usage with vscode's js debugger. It's a hack but until we have a more robust answer to debugging it's probably worthwhile.

It's worth noting that you can also directly use `--inspect`, but you have to do it with `--inspect=0` because each inflight will need its own port. However, now that we're using child processes, vscode's debugging is the only one I've gotten to attach correctly (because it inserts a `--require` hook to start the inspector instead)

Fixes #5863

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.