feat(sdk): `aws.QueueRef`
Introduce `aws.QueueRef` as a way to add an external queue reference to an application:

    bring aws;
    bring cloud;

    let myQueue = new aws.QueueRef("arn:aws:sqs:...."); // <-- ARN of an existing SQS queue
    new cloud.Function(inflight () => {
      myQueue.push("new message");
    });

This works both when running in the simulator and when running on AWS.

The Wing Console also includes some metadata about the queue such as the queue URL and the AWS Console URL:

### Misc

Added an option `link: bool` to `ui.FIeld` to indicate that the value returned is a link. We should make this a clickable link in the Wing Console.